backend/requirements.txt
fastapi
uvicorn[standard]
pandas
rank-bm25
python-dotenv
pydantic